Rework Bootstrapping to be able to reuse it for different spi implementations
